Fixed an issue with ID duplication, and also activated the correct display of messages in the UI

+++ src/main/java/com/foxminded/controllers/TeacherController.java
@@ -5,6 +5,7 @@ import com.foxminded.services.TeacherService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
@@ -14,27 +15,44 @@ public class TeacherController {
     @Autowired
     private TeacherService teacherService;
 
-    private static final String PAGE_TEACHER = "page-teacher";
     private static final String NAME_ATTRIBUTE_TEACHER = "teacher";
+    private static final String NAME_ATTRIBUTE_ALL = "allTeacher";
+    private static final String PAGE_UPDATE = "page-teacher-update";
+    private static final String PAGE_CREATE = "page-teacher-create";
+    private static final String PAGE_DELETE = "page-teacher-delete";
+    private static final String MESSAGE_INFO = "message";
+    private static final String HAS_ERRORS = "hasErrors";
+    private static final String SAVE_ALL = "savedSuccessful";
 
     @PostMapping("/create-teacher")
-    public String create(Model model, @ModelAttribute("teacher") Teacher teacher) {
-        model.addAttribute(NAME_ATTRIBUTE_TEACHER,new Teacher());
+    public String create(Model model, @ModelAttribute(NAME_ATTRIBUTE_TEACHER) Teacher teacher, BindingResult bindingResult) {
+        model.addAttribute(NAME_ATTRIBUTE_TEACHER, new Teacher());
+        boolean existenceCheck = teacherService.findAll().stream().anyMatch((teacherDTO -> teacherDTO.getFirstName().equals(teacher.getFirstName()) && (teacherDTO.getLastName().equals(teacher.getLastName()))));
+        if (bindingResult.hasErrors() || existenceCheck) {
+            model.addAttribute(HAS_ERRORS,true);
+            model.addAttribute(NAME_ATTRIBUTE_ALL, teacherService.findAll());
+            model.addAttribute(MESSAGE_INFO, "Such a teacher already exists");
+            return PAGE_CREATE;
+        }
         teacherService.create(teacher);
-        return PAGE_TEACHER;
+        model.addAttribute(SAVE_ALL,true);
+        model.addAttribute(NAME_ATTRIBUTE_ALL, teacherService.findAll());
+        model.addAttribute(MESSAGE_INFO, "Added teacher " + teacher);
+        return PAGE_CREATE;
     }
 
     @GetMapping("/create-teacher")
     public String getCreate(Model model) {
-        model.addAttribute(NAME_ATTRIBUTE_TEACHER,new Teacher());
-        return PAGE_TEACHER;
+        model.addAttribute(NAME_ATTRIBUTE_TEACHER, new Teacher());
+        model.addAttribute(NAME_ATTRIBUTE_ALL, teacherService.findAll());
+        return PAGE_CREATE;
     }
 
     @GetMapping("/all-teacher")
     public String findAll(Model model) {
-        model.addAttribute(NAME_ATTRIBUTE_TEACHER,new Teacher());
-        model.addAttribute("allTeachers",teacherService.findAll());
-        return PAGE_TEACHER;
+        model.addAttribute(NAME_ATTRIBUTE_TEACHER, new Teacher());
+        model.addAttribute(NAME_ATTRIBUTE_ALL, teacherService.findAll());
+        return PAGE_CREATE;
     }
 
     @PostMapping("/update-teacher")
@@ -43,27 +61,57 @@ public class TeacherController {
                          @RequestParam(required = false, name = "firstNameOld") String firstNameOld,
                          @RequestParam(required = false, name = "lastNameOld") String lastNameOld,
                          Model model) {
-        model.addAttribute(NAME_ATTRIBUTE_TEACHER,new Teacher());
-        teacherService.update(new Teacher(firstNameNew,lastNameNew), new Teacher(firstNameOld,lastNameOld));
-        return PAGE_TEACHER;
+        model.addAttribute(NAME_ATTRIBUTE_TEACHER, new Teacher());
+        Teacher teacherNew = new Teacher(firstNameNew, lastNameNew);
+        Teacher teacherOld = new Teacher(firstNameOld, lastNameOld);
+        boolean existenceCheckOld = teacherService.findAll().stream().anyMatch((teacherDTO -> teacherDTO.getFirstName().equals(firstNameOld) && (teacherDTO.getLastName().equals(lastNameOld))));
+        boolean existenceCheckNew = teacherService.findAll().stream().anyMatch((teacherDTO -> teacherDTO.getFirstName().equals(firstNameNew) && (teacherDTO.getLastName().equals(lastNameNew))));
+        try {
+            model.addAttribute(HAS_ERRORS,true);
+            if (!existenceCheckOld) {
+                model.addAttribute(MESSAGE_INFO, "Teacher with name " + teacherOld + " not found");
+                return PAGE_UPDATE;
+            } else if (existenceCheckNew) {
+                model.addAttribute(MESSAGE_INFO, "Teacher already " + teacherNew + " exists");
+                return PAGE_UPDATE;
+            }
+            teacherService.update(teacherNew, teacherOld);
+            model.addAttribute(SAVE_ALL,true);
+            model.addAttribute(MESSAGE_INFO, "Teacher changed from " + teacherOld + " to " + teacherNew);
+            return PAGE_UPDATE;
+        } finally {
+            model.addAttribute(NAME_ATTRIBUTE_ALL, teacherService.findAll());
+        }
     }
 
     @GetMapping("/update-teacher")
     public String getUpdate(Model model) {
-        model.addAttribute(NAME_ATTRIBUTE_TEACHER,new Teacher());
-        return PAGE_TEACHER;
+        model.addAttribute(NAME_ATTRIBUTE_TEACHER, new Teacher());
+        model.addAttribute(NAME_ATTRIBUTE_ALL, teacherService.findAll());
+        return PAGE_UPDATE;
     }
 
     @PostMapping("/delete-teacher")
-    public String delete(Model model, @ModelAttribute("teacher") Teacher teacher) {
-        model.addAttribute(NAME_ATTRIBUTE_TEACHER,new Teacher());
+    public String delete(Model model, @ModelAttribute(NAME_ATTRIBUTE_TEACHER) Teacher teacher, BindingResult bindingResult) {
+        model.addAttribute(NAME_ATTRIBUTE_TEACHER, new Teacher());
+        boolean existenceCheck = teacherService.findAll().stream().anyMatch((teacherDTO -> teacherDTO.getFirstName().equals(teacher.getFirstName()) && (teacherDTO.getLastName().equals(teacher.getLastName()))));
+        if (bindingResult.hasErrors() || !existenceCheck) {
+            model.addAttribute(HAS_ERRORS,true);
+            model.addAttribute(NAME_ATTRIBUTE_ALL, teacherService.findAll());
+            model.addAttribute(MESSAGE_INFO, "No such teacher found for deletion " + teacher);
+            return PAGE_DELETE;
+        }
         teacherService.delete(teacher);
-        return PAGE_TEACHER;
+        model.addAttribute(SAVE_ALL,true);
+        model.addAttribute(NAME_ATTRIBUTE_ALL, teacherService.findAll());
+        model.addAttribute(MESSAGE_INFO, "Deleted teacher " + teacher);
+        return PAGE_DELETE;
     }
 
     @GetMapping("/delete-teacher")
     public String getDelete(Model model) {
-        model.addAttribute(NAME_ATTRIBUTE_TEACHER,new Teacher());
-        return PAGE_TEACHER;
+        model.addAttribute(NAME_ATTRIBUTE_TEACHER, new Teacher());
+        model.addAttribute(NAME_ATTRIBUTE_ALL, teacherService.findAll());
+        return PAGE_DELETE;
     }
 }
